🎨 関数の記述場所を移動
